HBX-1910: Refactor JdbcMetadataBuilder into smaller responsibilities - Move class BinderMapping from 'org.hibernate.tool.internal.reveng' to 'org.hibernate.tool.internal.reveng.binder'
