+++ orm/src/main/java/org/hibernate/tool/internal/reveng/binder/BinderMapping.java
@@ -1,4 +1,4 @@
-package org.hibernate.tool.internal.reveng;
+package org.hibernate.tool.internal.reveng.binder;
 
 import org.hibernate.MappingException;
 import org.hibernate.boot.Metadata;
